+++ src/Makefile
@@ -1,4 +1,4 @@
 # Main code build rules
 
-src-y += canboot_main.c sched.c command.c flashcmd.c
+src-y += sched.c bootentry.c command.c flashcmd.c
 src-$(CONFIG_ENABLE_LED) += led.c
