--- src/canboot_main.c
@@ -1,126 +1,0 @@
-// Canboot main event loop
-//
-// Copyright (C) 2021 Eric Callahan <arksine.code@gmail.com>
-//
-// This file may be distributed under the terms of the GNU GPLv3 license.
-
-#include <string.h>         // strlen
-#include "autoconf.h"       // CONFIG_*
-#include "board/misc.h"     // udelay
-#include "board/flash.h"    // flash_read_block
-#include "board/gpio.h"     // gpio_in_setup
-#include "canboot_main.h"   // canboot_main
-#include "command.h"        // command_respond_ack
-#include "ctr.h"            // DECL_CTR
-#include "byteorder.h"      // cpu_to_le32
-#include "sched.h"          // sched_run_init
-
-#define REQUEST_SIG    0x5984E3FA6CA1589B // Random request sig
-
-static uint8_t complete;
-static uint32_t complete_endtime;
-
-void
-command_complete(uint32_t *data)
-{
-    uint32_t out[3];
-    command_respond_ack(CMD_COMPLETE, out, ARRAY_SIZE(out));
-    complete = 1;
-    complete_endtime = timer_read_time() + timer_from_us(100000);
-}
-
-void
-complete_task(void)
-{
-    if (complete && timer_is_before(complete_endtime, timer_read_time()))
-        jump_to_application();
-}
-DECL_TASK(complete_task);
-
-void
-command_connect(uint32_t *data)
-{
-    uint32_t mcuwords = DIV_ROUND_UP(strlen(CONFIG_MCU), 4);
-    uint32_t out[6 + mcuwords];
-    memset(out, 0, (6 + mcuwords) * 4);
-    out[2] = cpu_to_le32(PROTO_VERSION);
-    out[3] = cpu_to_le32(CONFIG_APPLICATION_START);
-    out[4] = cpu_to_le32(CONFIG_BLOCK_SIZE);
-    memcpy(&out[5], CONFIG_MCU, strlen(CONFIG_MCU));
-    command_respond_ack(CMD_CONNECT, out, ARRAY_SIZE(out));
-}
-
-static inline uint8_t
-check_application_code(void)
-{
-    // Read the first block of memory, if it
-    // is all 0xFF then no application has been flashed
-    uint8_t buf[CONFIG_BLOCK_SIZE];
-    flash_read_block(CONFIG_APPLICATION_START, (uint32_t*)buf);
-    for (uint8_t i = 0; i < CONFIG_BLOCK_SIZE; i++) {
-        if (buf[i] != 0xFF)
-            return 1;
-    }
-    return 0;
-}
-
-// Generated by buildcommands.py
-DECL_CTR("DECL_BUTTON " __stringify(CONFIG_BUTTON_PIN));
-extern int32_t button_gpio, button_high, button_pullup;
-
-// Check for a bootloader request via double tap of reset button
-static int
-check_button_pressed(void)
-{
-    if (!CONFIG_ENABLE_BUTTON)
-        return 0;
-    struct gpio_in button = gpio_in_setup(button_gpio, button_pullup);
-    udelay(10);
-    return gpio_in_read(button) == button_high;
-}
-
-// Check for a bootloader request via double tap of reset button
-static void
-check_double_reset(void)
-{
-    if (!CONFIG_ENABLE_DOUBLE_RESET)
-        return;
-    // set request signature and delay for two seconds.  This enters the bootloader if
-    // the reset button is double clicked
-    set_bootup_code(REQUEST_SIG);
-    udelay(500000);
-    set_bootup_code(0);
-    // No reset, read the key back out to clear it
-}
-
-
-/****************************************************************
- * Startup
- ****************************************************************/
-static void
-enter_bootloader(void)
-{
-    sched_run_init();
-
-    for (;;)
-        sched_run_tasks();
-}
-
-// Main loop of program
-void
-canboot_main(void)
-{
-    // Enter the bootloader in the following conditions:
-    // - The request signature is set in memory (request from app)
-    // - No application code is present
-    uint64_t bootup_code = get_bootup_code();
-    if (bootup_code == REQUEST_SIG || !check_application_code()
-        || check_button_pressed()) {
-        set_bootup_code(0);
-        enter_bootloader();
-    }
-    check_double_reset();
-
-    // jump to app
-    jump_to_application();
-}
