--- src/canboot_main.h
@@ -1,6 +1,0 @@
-#ifndef __CANBOOT_MAIN_H
-#define __CANBOOT_MAIN_H
-
-void canboot_main(void);
-
-#endif // canboot_main.h
